add user id to bookmark

+++ FolkTell/FolkTell/src/dao/BookmarkDao.cpp
@@ -11,7 +11,7 @@ BookmarkDao& BookmarkDao::getDao(){
 }
 bool BookmarkDao::createTable(){
     QString cmd = "CREATE TABLE IF NOT EXISTS " + this->getTableName() +
-                " (ID INTEGER PRIMARY KEY AUTOINCREMENT, GID INTEGER, NAME TEXT, URL TEXT, ICON TEXT);";
+                " (UID INTEGER NOT NULL, ID INTEGER NOT NULL, GID INTEGER, NAME TEXT UNIQUE, URL TEXT, ICON TEXT, PRIMARY KEY(UID, ID));";
     QSqlQuery query(this->db);
     bool ok = query.exec(cmd);
 
@@ -24,26 +24,26 @@ bool BookmarkDao::createTable(){
 }
 
 
-QVector<QVariant> BookmarkDao::QueryById(const int& id){
-    QString cmd = "SELECT * from " + this->getTableName() + " WHERE ID="+QString::number(id);
+QVector<QVariant> BookmarkDao::QueryByUidAndId(const int& uid, const int& id){
+    QString cmd = "SELECT * from " + this->getTableName() + " WHERE ID="+QString::number(id)+ " AND UID="+QString::number(uid);
     QSqlQuery query(this->db);
     query.prepare(cmd);
     
     QVector<QVariant> ret;
     if(!query.exec()){
-        qDebug() << "[error] fail to query by id,  " << query.lastError().text();
+        qDebug() << "[error] fail to query by id and uid,  " << query.lastError().text();
         return ret;
     }
     
     while(query.next()){
-        ret.append({query.value(0), query.value(1),query.value(2),query.value(3),query.value(4)});
+        ret.append({query.value(0), query.value(1),query.value(2),query.value(3),query.value(4), query.value(5)});
     }
 
     return ret;
 }
 
-QVector<QVector<QVariant>> BookmarkDao::QueryByGroupId(const int& gid){
-    QString cmd = "SELECT * from " + this->getTableName() + " WHERE GID="+QString::number(gid);
+QVector<QVector<QVariant>> BookmarkDao::QueryByUidAndGroupId(const int& uid, const int& gid){
+    QString cmd = "SELECT * from " + this->getTableName() + " WHERE GID="+QString::number(gid) + " AND UID="+QString::number(uid);
     QSqlQuery query(this->db);
     query.prepare(cmd);
 
@@ -54,18 +54,18 @@ QVector<QVector<QVariant>> BookmarkDao::QueryByGroupId(const int& gid){
     }
 
     while(query.next()){
-        QVector<QVariant> temp = {query.value(0), query.value(1),query.value(2),query.value(3),query.value(4)};
+        QVector<QVariant> temp = {query.value(0), query.value(1),query.value(2),query.value(3),query.value(4), query.value(5)};
         ret.emplace_back(temp);
     }
 
     return ret;
 }
 
-QVector<QVector<QVariant>> BookmarkDao::QueryByName(const QString& name){
-    QString cmd = "SELECT * from " + this->getTableName() + " WHERE NAME="+name;
+QVector<QVector<QVariant>> BookmarkDao::QueryByUidAndName(const int& uid, const QString& name){
+    QString cmd = "SELECT * from " + this->getTableName() + " WHERE NAME="+name + " AND UID="+QString::number(uid);
     QSqlQuery query(this->db);
     query.prepare(cmd);
-
+    qDebug() << this->getcolumns();
     QVector<QVector<QVariant>> ret;
     if(!query.exec()){
         qDebug() << "[error] fail to query by name,  " << query.lastError().text();
@@ -73,15 +73,15 @@ QVector<QVector<QVariant>> BookmarkDao::QueryByName(const QString& name){
     }
 
     while(query.next()){
-        QVector<QVariant> temp = {query.value(0), query.value(1),query.value(2),query.value(3),query.value(4)};
+        QVector<QVariant> temp = {query.value(0), query.value(1),query.value(2),query.value(3),query.value(4), query.value(5)};
         ret.emplace_back(temp);
     }
 
     return ret;
 }
 
-bool BookmarkDao::setName(const int& id, const QString& name){
-    QString cmd = "UPDATE " + this->table_name + "SET NAME=:name WHERE ID=:id";
+bool BookmarkDao::setName(const int& uid, const int& id, const QString& name){
+    QString cmd = "UPDATE " + this->table_name + "SET NAME=:name WHERE ID=:id AND UID=:uid";
     QSqlQuery query(this->db);
     
     if(!query.prepare(cmd)){
@@ -91,6 +91,7 @@ bool BookmarkDao::setName(const int& id, const QString& name){
 
     query.bindValue(":name", name);
     query.bindValue(":id", id);
+    query.bindValue(":uid", uid);
 
     if(!query.exec()){
         qDebug() << "[error] fail to exec cmd in setname: " << query.lastError().text();
@@ -100,8 +101,8 @@ bool BookmarkDao::setName(const int& id, const QString& name){
     return true;
 }
 
-bool BookmarkDao::setGid(const int& id, const int& gid){
-    QString cmd = "UPDATE " + this->table_name + "SET GID=:gid WHERE ID=:id";
+bool BookmarkDao::setGid(const int& uid, const int& id, const int& gid){
+    QString cmd = "UPDATE " + this->table_name + "SET GID=:gid WHERE ID=:id AND UID=:uid";
     QSqlQuery query(this->db);
     
     if(!query.prepare(cmd)){
@@ -111,6 +112,7 @@ bool BookmarkDao::setGid(const int& id, const int& gid){
 
     query.bindValue(":gid", gid);
     query.bindValue(":id", id);
+    query.bindValue(":uid", uid);
 
     if(!query.exec()){
         qDebug() << "[error] fail to exec cmd in setgid: " << query.lastError().text();
@@ -120,8 +122,8 @@ bool BookmarkDao::setGid(const int& id, const int& gid){
     return true;
 }
 
-bool BookmarkDao::setIcon(const int& id, const QString& icon){
-    QString cmd = "UPDATE " + this->table_name + "SET ICON=:icon WHERE ID=:id";
+bool BookmarkDao::setIcon(const int& uid, const int& id, const QString& icon){
+    QString cmd = "UPDATE " + this->table_name + "SET ICON=:icon WHERE ID=:id AND UID=:uid";
     QSqlQuery query(this->db);
     
     if(!query.prepare(cmd)){
@@ -131,6 +133,7 @@ bool BookmarkDao::setIcon(const int& id, const QString& icon){
 
     query.bindValue(":icon", icon);
     query.bindValue(":id", id);
+    query.bindValue(":uid", uid);
 
     if(!query.exec()){
         qDebug() << "[error] fail to exec cmd in seticon: " << query.lastError().text();
@@ -140,8 +143,8 @@ bool BookmarkDao::setIcon(const int& id, const QString& icon){
     return true;
 }
 
-bool BookmarkDao::setUrl(const int& id, const QUrl& url){
-    QString cmd = "UPDATE " + this->table_name + "SET URL=:url WHERE ID=:id";
+bool BookmarkDao::setUrl(const int& uid, const int& id, const QUrl& url){
+    QString cmd = "UPDATE " + this->table_name + "SET URL=:url WHERE ID=:id AND UID=:uid";
     QSqlQuery query(this->db);
     
     if(!query.prepare(cmd)){
@@ -151,6 +154,7 @@ bool BookmarkDao::setUrl(const int& id, const QUrl& url){
 
     query.bindValue(":url", url);
     query.bindValue(":id", id);
+    query.bindValue(":uid", uid);
 
     if(!query.exec()){
         qDebug() << "[error] fail to exec cmd in seturl: " << query.lastError().text();
@@ -161,11 +165,15 @@ bool BookmarkDao::setUrl(const int& id, const QUrl& url){
 }
 
 
-bool BookmarkDao::insert(const int& gid, const QString& name, const QUrl& url){
-    QString cmd = "INSERT INTO " + this->getTableName() + "(GID, NAME, URL)"+" VALUES(:gid, :name, :url)";
+bool BookmarkDao::insert(const int& uid, const int& gid, const QString& name, const QUrl& url, const QUrl& icon){
+    QString cmd = "INSERT INTO " + this->getTableName() + "(UID, ID, GID, NAME, URL)"+" VALUES(:uid, :id, :gid, :name, :url, :icon)";
 
     QSqlQuery query(this->db);
     query.prepare(cmd);
+    int id = qHash(name);
+
+    query.bindValue(":uid", uid);
+    query.bindValue(":id", id);
     query.bindValue(":gid", gid);
     query.bindValue(":name", name);
     query.bindValue(":url", url);
@@ -178,12 +186,13 @@ bool BookmarkDao::insert(const int& gid, const QString& name, const QUrl& url){
     return true;
 }
 
-bool BookmarkDao::remove(const int& id){
-    QString cmd = "DELETE FROM " + this->table_name + " WHERE ID=:id";
+bool BookmarkDao::remove(const int& uid, const int& id){
+    QString cmd = "DELETE FROM " + this->table_name + " WHERE ID=:id AND UID=:uid";
 
     QSqlQuery query(this->db);
     query.prepare(cmd);
     query.bindValue(":id", id);
+    query.bindValue(":uid", uid);
 
     if(!query.exec()){
         qDebug() << "[error] fail to delete,  " << query.lastError().text();
@@ -193,6 +202,7 @@ bool BookmarkDao::remove(const int& id){
     return true;
 }
 
+/*
 QVector<QVariant> BookmarkDao::getcolumns(){
     QString cmd = "PRAGMA  table_info("+this->table_name+")";
     QSqlQuery query(this->db);
@@ -210,7 +220,7 @@ QVector<QVariant> BookmarkDao::getcolumns(){
 
     return ret;
 }
-
+*/
 bool BookmarkDao::deleteTable(){
     QString cmd = "DROP TABLE " + this->table_name;
